# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Campus RPG API 문서
  version: 0.1.0
  description: |
    현재 백엔드에 구현된 API를 기준으로 작성한 OpenAPI 초안입니다.
    `backend/server.js`와 각 라우트 모듈의 실제 경로를 반영합니다.
servers:
  - url: http://localhost:8888
    description: 로컬 백엔드 서버
...
paths:
...
  /api/announcements:
    get:
      tags: [announcements]
      summary: 공지 목록
      description: |
        최신순(`createdAt`) 공지 요약 목록을 반환합니다. 응답 키는 camelCase(`createdAt`)입니다.
      responses:
        "200":
          description: 공지 목록
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/AnnouncementListItem"
        "500":
          description: 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/announcements/{id}:
    get:
      tags: [announcements]
      summary: 공지 상세
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 공지 본문 포함
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnnouncementDetail"
        "400":
          description: 잘못된 ID
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 공지 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
  /api/friends:
    get:
      tags: [friends]
      summary: 친구 목록 조회
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 친구 목록
          content:
            application/json:
              schema:
                type: object
                required: [friends]
                properties:
                  friends:
                    type: array
                    items:
                      $ref: "#/components/schemas/FriendItem"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 친구 목록 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/friends/order:
    patch:
      tags: [friends]
      summary: 친구 정렬 순서 저장
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [orderedUserIds]
              properties:
                orderedUserIds:
                  type: array
                  items:
                    type: integer
                    minimum: 1
      responses:
        "200":
          description: 저장 성공
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 친구 정렬 순서 저장 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/friends/{friendUserId}:
    delete:
      tags: [friends]
      summary: 친구 삭제
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: friendUserId
          required: true
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 삭제 성공
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 친구 관계 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 친구 삭제 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/friends/requests:
    post:
      tags: [friends]
      summary: 친구 코드로 친구 요청
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [friendCode]
              properties:
                friendCode:
                  type: string
                friend_code:
                  type: string
                  description: friendCode 대신 사용할 수 있는 대체 필드
      responses:
        "201":
          description: 요청 생성 성공
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 친구 코드 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 중복 요청, 이미 친구, 또는 자기 자신에게 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 친구 요청 생성 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/friends/requests/incoming:
    get:
      tags: [friends]
      summary: 받은 친구 요청 조회
      security:
        - bearerAuth: []
      responses:
        "200":
          description: 받은 요청 목록
          content:
            application/json:
              schema:
                type: object
                required: [requests]
                properties:
                  requests:
                    type: array
                    items:
                      $ref: "#/components/schemas/FriendRequestItem"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 받은 친구 요청 조회 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/friends/requests/{requestId}/accept:
    post:
      tags: [friends]
      summary: 친구 요청 수락
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          description: friend_requests 테이블의 PK(id)
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 수락 성공
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 요청 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 이미 처리된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 친구 요청 수락 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
  /api/friends/requests/{requestId}/reject:
    post:
      tags: [friends]
      summary: 친구 요청 거절
      security:
        - bearerAuth: []
      parameters:
        - in: path
          name: requestId
          required: true
          description: friend_requests 테이블의 PK(id)
          schema:
            type: integer
            minimum: 1
      responses:
        "200":
          description: 거절 성공
        "400":
          description: 잘못된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "401":
          description: 인증 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: 요청 없음
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "409":
          description: 이미 처리된 요청
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "500":
          description: 친구 요청 거절 실패
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        message:
          type: string
          description: 사람이 읽기 쉬운 부가 설명(선택)
...
    AnnouncementListItem:
      type: object
      required: [id, title, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        createdAt:
          type: string
    AnnouncementDetail:
      type: object
      required: [id, title, content, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        content:
          type: string
        createdAt:
          type: string
...
    FriendItem:
      type: object
      required: [friendUserId, sortOrder, nickname]
      properties:
        friendUserId:
          type: integer
          example: 2
        sortOrder:
          type: integer
          example: 0
        nickname:
          type: string
          example: campusMate
        intro:
          type: string
          example: 안녕하세요
        avatar:
          type: string
          example: "🐶"
    FriendRequestItem:
      type: object
      required: [id, fromUserId, nickname]
      properties:
        id:
          type: integer
          example: 10
        fromUserId:
          type: integer
          example: 3
        createdAt:
          type: string
          example: "2026-05-06 15:22:10"
        nickname:
          type: string
          example: requester
        intro:
          type: string
          example: 친하게 지내요
        avatar:
          type: string
          example: "🦊"
...